React router configured
